requirements.txt
fastapi
uvicorn[standard]
PyPDF2
sentence-transformers
faiss-cpu
langchain[text-splitters]
openai
python-dotenv
tqdm
transformers
torch
